Fixed some visual bugs, added animations and color to stickman, cursor not showing

+++ LPOORTAL_Game_Server/core/src/view/entities/TextureManager.java
@@ -2,7 +2,6 @@ package view.entities;
 
 import java.util.HashMap;
 
-import com.badlogic.gdx.assets.AssetManager;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Animation;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
@@ -13,7 +12,6 @@ import model.entities.StickmanModel.Stickman_Animation;
 import model.entities.StickmanModel.Stickman_Facing_Direction;
 
 public class TextureManager {
-	
 	
 	private HashMap<Pair<Stickman_Animation, Stickman_Facing_Direction>, Animation<TextureRegion>> stickmanAnimations;
 	private Texture background, cursor, lineTexture;
@@ -35,13 +33,14 @@ public class TextureManager {
 	}
 
 	private void populateAnimations() {
-		addPlayerAnimation("Idle.png", Stickman_Animation.IDLE, (int) (4f/30f));
-		addPlayerAnimation("Lift_Off.png", Stickman_Animation.LIFT_OFF, (int) (2f/30f));
-		addPlayerAnimation("Walking.png", Stickman_Animation.WALKING, (int) (3f/30f));
-		addPlayerAnimation("mid_flight.png", Stickman_Animation.JUMPING, (int) (30f/30f));
+		addPlayerAnimation("Idle.png", Stickman_Animation.IDLE, 4f/30f);
+		addPlayerAnimation("Lift_Off.png", Stickman_Animation.LIFT_OFF, 2f/30f);
+		addPlayerAnimation("Walking.png", Stickman_Animation.WALKING, 3f/30f);
+		addPlayerAnimation("mid_flight.png", Stickman_Animation.JUMPING, 30f/30f);
 	}
-
-	private void addPlayerAnimation(String source, Stickman_Animation animName, int frameTime) {
+	
+	
+	private void addPlayerAnimation(String source, Stickman_Animation animName, float frameTime) {
 		
 		Texture t = new Texture(source);
 		Texture tReversed = new Texture(source);
@@ -59,6 +58,7 @@ public class TextureManager {
 		
 		TextureRegion[] framesReversed = new TextureRegion[trReversed[0].length];
 		System.arraycopy(trReversed[0], 0, framesReversed, 0, trReversed[0].length);
+		
 		
 	    stickmanAnimations.put(new Pair<Stickman_Animation, Stickman_Facing_Direction>(
 	    		animName, Stickman_Facing_Direction.RIGHT),
